editor updates
