+++ lingo3d/src/events/onEditorLayout.ts
@@ -1,0 +1,3 @@
+import { event } from "@lincode/events"
+
+export const [emitEditorLayout, onEditorLayout] = event()
