+++ lingo3d/src/states/useEditorMounted.ts
@@ -1,8 +1,3 @@
 import store from "@lincode/reactivity"
 
-export const [setEditorMounted, getEditorMounted] = store(0)
-
-export const increaseEditorMounted = () =>
-    setEditorMounted(getEditorMounted() + 1)
-export const decreaseEditorMounted = () =>
-    setEditorMounted(getEditorMounted() - 1)
+export const [setEditorMounted, getEditorMounted] = store(false)
