--- lingo3d/src/states/useNodeEditor.ts
@@ -1,3 +1,0 @@
-import store from "@lincode/reactivity"
-
-export const [setNodeEditor, getNodeEditor] = store(false)
